Able to use project components as golang library
