+++ cmd/rollback.go
@@ -69,7 +69,7 @@ func rollbackCmd() *cobra.Command {
 
 func (d *rollback) backcastHelm3() error {
 	namespace := os.Getenv("HELM_NAMESPACE")
-	excludes := []string{helm3TestHook, helm2TestSuccessHook}
+	excludes := []string{manifest.Helm3TestHook, manifest.Helm2TestSuccessHook}
 	if d.includeTests {
 		excludes = []string{}
 	}
